выделил основное дз в отдельный пакет mainTask, и еще обработаны случаи отключения клиента от сервера.
