+++ Lesson6.WorkWithNet/src/mainTask/EchoServer.java
@@ -1,3 +1,5 @@
+package mainTask;
+
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.net.ServerSocket;
